Split out Glyph stories
